@@ -8,20 +8,20 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3",
         "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^6.5",
-        "laravel/framework": "^7.25",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.4"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.7",
         "fzaninotto/faker": "^1.9.1",
-        "mockery/mockery": "^1.4.1",
-        "nunomaduro/collision": "^4.2",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
         "phpunit/phpunit": "^9.3",
-		"barryvdh/laravel-debugbar": "^3.4"
+		"barryvdh/laravel-debugbar": "^3.5"
     },
     "config": {
         "optimize-autoloader": true,
@@ -35,12 +35,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
