@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,4 +14,4 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/user/{id}', 'UserController@show')->where('id', '[0-9]+');
+Route::get('/user/{id}', [UserController::class, 'show'])->where('id', '[0-9]+');
